Factor all the metadata panel rows into single metadata_rows file.
