+++ client/scripts/views/components/admin_panel/chain_metadata_management_table.ts
@@ -1,9 +1,9 @@
 import m from 'mithril';
 import { ChainNetwork } from 'client/scripts/models';
 import { Button, Table } from 'construct-ui';
-import InputPropertyRow from './input_property_row';
-import ManageRolesRow from './manage_roles_row';
+
 import { IChainOrCommMetadataManagementAttrs } from './community_metadata_management_table';
+import { InputPropertyRow, ManageRolesRow } from './metadata_rows';
 
 interface IChainMetadataManagementState {
   name: string;
